Work around swapped Joy-Con button diamond origins
NB: as a side effect, if Nintendo button layout is switched off in Steam's
settings, paired Joy-Cons now display the physical button names in-game:
"B" for action, "A" for cancel, etc. This may or may not be desired by the
player. For the Switch Pro Controller, the game still displays the logical
button names in an analogous situation: "A" for action, "B" for cancel.

There's not much I can do to remove this discrepancy, as it's caused by
Steam Input itself reporting opposite button IDs (action origins) for the
Switch Pro Controller and for the Joy-Cons but reusing the same set of
origins for both. I guess I could split off "Xbox-layout paired Joy-Cons"
into its own type of controller, but then I'd need to detect it based on
the origins, which seems fragile: what would I do if Steam were to claim
some actions come from Xbox360 origins and some others from Switch ones?
And I have, in fact, seen it report mixed kinds of origins before,
although not in this particular instance (yet).