More sensible exploration prompts for lone Joy-Con
The default exploration prompt for a lone Joy-Con was an "A"
until commit 9b38c0a0dd1aa042b20cb06ec3af027254550646,
which flipped lone Joy-Con mappings to better accommodate
Xbox360 origins in Steam's "Nintendo layout off" mode,
thereby changing the default exploration prompt to "B"
when Nintendo layout is *on* and (from "B") to "A" when
it's off. The exploration prompt change was actually
an oversight, and showing "B" for the Nintendo layout
doesn't really make sense.

To suit both layout modes *and* to resolve the oddness that letters are
shown at all despite other prompts showing button diamond directions,
change the exploration prompts for a lone Joy-Con to depict the
physical button of the rotated right Joy-Con in the corresponding
position, i. e. the physical button that needs to be pressed.
If the actual Joy-Con is a left one, then this physical meaning
is lost, but it still seems better than any alternative.